+++ src/SelectedWallet/WalletSelection/WalletSelectionScreen.tsx
@@ -1,14 +1,12 @@
-import {useFocusEffect, useNavigation} from '@react-navigation/native'
-import {delay} from 'bluebird'
+import {useNavigation} from '@react-navigation/native'
 import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
-import {Alert, FlatList, Linking, RefreshControl, StyleSheet, Text, TouchableOpacity} from 'react-native'
+import {FlatList, Linking, RefreshControl, StyleSheet, Text, TouchableOpacity} from 'react-native'
 import {SafeAreaView} from 'react-native-safe-area-context'
-import {useMutation, UseMutationOptions, useQueryClient} from 'react-query'
 
 import {useAuth} from '../../auth/AuthProvider'
 import {Button, Icon, PleaseWaitModal, StatusBar} from '../../components'
-import {useCloseWallet, useWalletMetas} from '../../hooks'
+import {useCloseWallet, useOpenWallet, useWalletMetas} from '../../hooks'
 import globalMessages, {errorMessages} from '../../i18n/global-messages'
 import {showErrorDialog} from '../../legacy/actions'
 import {CONFIG, isNightly} from '../../legacy/config'
@@ -18,7 +16,7 @@ import {WalletMeta} from '../../legacy/state'
 import {useWalletNavigation} from '../../navigation'
 import {COLORS} from '../../theme'
 import {useWalletManager} from '../../WalletManager'
-import {KeysAreInvalid, SystemAuthDisabled, walletManager, YoroiWallet} from '../../yoroi-wallets'
+import {KeysAreInvalid, SystemAuthDisabled} from '../../yoroi-wallets'
 import {useSetSelectedWallet, useSetSelectedWalletMeta} from '..'
 import {WalletListItem} from './WalletListItem'
 
@@ -32,34 +30,31 @@ export const WalletSelectionScreen = () => {
   const selectWalletMeta = useSetSelectedWalletMeta()
   const selectWallet = useSetSelectedWallet()
   const intl = useIntl()
-  const queryClient = useQueryClient()
+
   const {logout} = useAuth()
 
   const {closeWallet} = useCloseWallet()
 
-  useFocusEffect(closeWallet)
-
   const {openWallet, isLoading} = useOpenWallet({
-    onSuccess: ({wallet, walletMeta}) => {
+    onSuccess: ([wallet, walletMeta]) => {
       selectWalletMeta(walletMeta)
       selectWallet(wallet)
-      wallet.subscribeOnTxHistoryUpdate(() => queryClient.invalidateQueries([wallet.id, 'lockedAmount']))
+
       navigateToTxHistory()
     },
     onError: async (error) => {
+      closeWallet()
+
       if (error instanceof SystemAuthDisabled) {
-        closeWallet()
         await showErrorDialog(errorMessages.enableSystemAuthFirst, intl)
         resetToWalletSelection()
       } else if (error instanceof InvalidState) {
-        closeWallet()
         await showErrorDialog(errorMessages.walletStateInvalid, intl)
       } else if (error instanceof KeysAreInvalid) {
         await showErrorDialog(errorMessages.walletKeysInvalidated, intl)
-        closeWallet()
         logout()
       } else {
-        Alert.alert('QWEQWE', error.message)
+        await showErrorDialog(errorMessages.generalError, intl, {message: error.message})
       }
     },
   })
@@ -229,31 +224,6 @@ const OnlyDevButton = () => {
   if (!__DEV__) return null
 
   return <Button onPress={() => navigation.navigate('developer')} title="Dev options" style={styles.button} />
-}
-
-const useOpenWallet = (
-  options?: UseMutationOptions<{wallet: YoroiWallet; walletMeta: WalletMeta}, Error, WalletMeta>,
-) => {
-  const {closeWallet} = useCloseWallet()
-
-  const mutation = useMutation({
-    ...options,
-    mutationFn: async (walletMeta) => {
-      closeWallet()
-      await delay(500)
-      const [newWallet, newWalletMeta] = await walletManager.openWallet(walletMeta)
-
-      return {
-        wallet: newWallet,
-        walletMeta: newWalletMeta,
-      }
-    },
-  })
-
-  return {
-    openWallet: mutation.mutate,
-    ...mutation,
-  }
 }
 
 const styles = StyleSheet.create({
