+++ src/yoroi-wallets/walletManager.ts
@@ -300,6 +300,7 @@ export class WalletManager {
   }
 
   async openWallet(walletMeta: WalletMeta): Promise<[YoroiWallet, WalletMeta]> {
+    await this.closeWallet()
     assert.preconditionCheck(!!walletMeta.id, 'openWallet:: !!id')
     const data = await storage.read(`/wallet/${walletMeta.id}/data`)
     const appSettings = await readAppSettings()
